don't delete category has products and don't add the same name

+++ Web.DataAccess/Repositories/CategoryRepository.cs
@@ -19,6 +19,10 @@ namespace Web.DataAccess.Repositories
             if (validationResult is not null)
                 return validationResult;
 
+            var existingCategory = await GetByAsync(x => x.Name == categoryVM.Name);
+            if (existingCategory != null)
+                return new List<ValidationError> { new("Duplicate Category", "Category with this name already exists") };
+
             var category = categoryVM.Adapt<Category>();
             category.ImageName = await SaveImageAsync(categoryVM.Image);
             await _context.Categories.AddAsync(category, cancellationToken);
@@ -84,6 +88,10 @@ namespace Web.DataAccess.Repositories
             var categoryFromDb = await GetByAsync(x => x.Id == id);
             if (categoryFromDb == null)
                 return new List<ValidationError> { new("Not Found", "Category not found") };
+
+            var existingProducts = await _context.Products.AnyAsync(x => x.CategoryId == id);
+            if (existingProducts)
+                return new List<ValidationError> { new("Cannot Delete", "Category cannot be deleted as it has associated products") };
 
             DeleteImageFile(categoryFromDb.ImageName);
             _context.Categories.Remove(categoryFromDb);
